+++ app/pb_kits/playbook/pb_radio/docs/_radio_error.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import { Radio } from '../..'
+
+const RadioError = () => {
+  return (
+    <div>
+      <Radio
+          error
+          label="Power"
+          name="Group2"
+          value="Power"
+      />
+    </div>
+  )
+}
+export default RadioError