+++ spec/pb_kits/playbook/kits/radio_spec.rb
@@ -8,6 +8,7 @@ RSpec.describe Playbook::PbRadio::Radio do
   it { is_expected.to define_partial }
   it { is_expected.to define_boolean_prop(:checked).with_default(false) }
   it { is_expected.to define_boolean_prop(:dark).with_default(false) }
+  it { is_expected.to define_boolean_prop(:error).with_default(false) }
   it { is_expected.to define_prop(:name) }
   it { is_expected.to define_prop(:value) }
 
@@ -18,6 +19,8 @@ RSpec.describe Playbook::PbRadio::Radio do
         expect(subject.new(dark: true).classname).to eq "pb_radio_kit_dark"
         expect(subject.new(checked: true).classname).to eq "pb_radio_kit"
         expect(subject.new(dark: true, checked: true).classname).to eq "pb_radio_kit_dark"
+        expect(subject.new(checked: false, error: true).classname).to eq "pb_radio_kit error"
+        expect(subject.new(checked: false, dark: true, error: true).classname).to eq "pb_radio_kit_dark error"
     end
   end
 end